+++ CHANGELOG.md
@@ -112,7 +112,7 @@
 ## 0.2.3
 
 ##### Added
-  - HasImages trait for Eloquent models (provides `hasMany Image` relationship)
+  - HasUploads trait for Eloquent models (provides `hasMany Image` relationship)
 
 ##### Fixed
   - Pages module tests